moodlens/requirements.txt
# MoodLens - AI Mental Health Journal
# Python 3.13 compatible dependencies
# Install: pip install -r requirements.txt

# ── Web Framework ─────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ── Database ──────────────────────────────────────────────────
sqlalchemy==2.0.36
aiosqlite==0.20.0            # async SQLite driver

# ── Validation & Serialisation ───────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0

# ── Authentication & Security ─────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# ── Caching ───────────────────────────────────────────────────
cachetools==5.5.0

# ── AI / Claude API ───────────────────────────────────────────
anthropic==0.43.0

# ── Environment & Config ──────────────────────────────────────
python-dotenv==1.0.1

# ── Utilities ─────────────────────────────────────────────────
httpx==0.28.1
tenacity==9.0.0              # retry logic for external API calls
greenlet>=3.0.0
